refactore: split constructs
