+++ test/langfuse-with-aws-cdk-stg.test.ts
@@ -41,7 +41,6 @@ test('snapshot test for stg', () => {
 
     disableEmailPasswordAuth: appConfig.disableEmailPasswordAuth,
     certificateForCognito: usEast1Stack?.certificateForCognito,
-    enableCognitoAuth: appConfig.enableCognitoAuth,
   });
 
   // THEN
